Started working on deployment procedures. A bit of reshuffling to make code tidier.
