+++ contracts/package.json
@@ -11,11 +11,14 @@
   "license": "ISC",
   "devDependencies": {
     "@nomicfoundation/hardhat-toolbox": "^2.0.0",
+    "@nomiclabs/hardhat-ethers": "npm:hardhat-deploy-ethers@^0.3.0-beta.13",
     "@openzeppelin/contracts": "^4.5.0",
     "@openzeppelin/hardhat-upgrades": "^1.21.0",
     "@solidstate/hardhat-bytecode-exporter": "^1.1.1",
+    "ethers": "^5.7.2",
     "hardhat": "^2.12.4",
     "hardhat-abi-exporter": "^2.10.1",
+    "hardhat-deploy": "^0.11.22",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.4"
   }
